Move test file system into server to fix linking error on macOS
For now just move it unmodified and remove the scaffolding for setting
it from outside the server library.
